% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{EM Function}
\usage{
EM(x0, k)
}
\arguments{
\item{x0}{input data (vector)}

\item{k}{amount of clusters (mixture components)}
}
\value{
Parameters for each distribution
}
\description{
This function realized EM algorithm (Expectation-Maximization algorithm)
for data clustering. In this case, for mixture models.
}
\details{
In initial step (Step_0) explorer must determine the initial gaussian mixture model 
parameters. Explorer can take any random parameters or (using surface analysis and some assumptions) 
take anothers.

In the first step (called E_Step) with function E_Step explorer calculate 
the posterior probabilities (which named "posterior.df") for each 
item of initial dataset.

In the second step (called M_Step) with function M_Step explorer
update component parameters (using likelihood function).

After that, explorer repeat M_Step (100 times or until difference between current value of 
logarithm of likelihood function and previous value of logarithm of likelihood function
will be less than 10^(-6) 
in other words: logarithm of likelihood function didn’t change much). 

In the end explorer has updated probabilities for each item and
updated parameters for each distribution (explorer should look at E.Step and M.Step)
}
\author{
hdrbv
}
