% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_em.R
\name{plot_em}
\alias{plot_em}
\title{plot_em Function}
\usage{
plot_em(x0, EM)
}
\arguments{
\item{x0}{input data (vector)}

\item{EM}{application of EM function to input data}
}
\value{
Plot of mixture
}
\description{
This function created plot of mixture distributions as a result of separation process via EM algorithm (Expectation-Maximization algorithm)
}
\examples{
For example we want to separate 2 Gaussian distributions, 
estimate parameters of each one and plot distributions.
Let us assume that vector x1 - mixture of these distributions. 
So we can use EM algorithm here:   
EM1 <- sepro::EM(x0 = x0, k = 2).
And create plot:
plot_em(x0, EM1)
}
\author{
hdrbv
}
