\name{sepro-package}
\alias{sepro-package}
\alias{sepro}
\docType{package}
\title{
\packageTitle{sepro}
}
\description{
\packageDescription{sepro}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{sepro}
EM function from package Separation Process (sepro) - 
realisation of EM algorithm (Expectation–maximization algorithm)
}
\author{
\packageAuthor{sepro}
}
\references{
Dellaert, Frank (2002). "The Expectation Maximization Algorithm"
}
\examples{
"For example we want to separate 2 Gaussian distributions  
and estimate parameters of each one. 
Let us assume that vector x1 - mixture of these distributions. 
So we can use EM algorithm here:
EM(x1, k = 2)"
}
