\name{sepro-package}
\alias{sepro-package}
\alias{sepro}
\docType{package}
\title{
\packageTitle{sepro}
}
\description{
\packageDescription{sepro}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{sepro}
EM function from package Separation Process (sepro) - 
realisation of EM algorithm (Expectation–maximization algorithm)
}
\author{
\packageAuthor{sepro}
}
\references{
Dellaert, Frank (2002). "The Expectation Maximization Algorithm"
}
